feat : all feature implement - status ok
