+++ back-end/src/controllers/postController.ts
@@ -18,12 +18,14 @@ async function publishPost(req: Request, res: Response) {
 }
 
 async function getPosts(req: Request, res: Response) {
-  const posts = await postService.getBestPosts();
+  const userId = res.locals.user.id;
+  const posts = await postService.getBestPosts(userId);
   res.json(posts);
 }
 
 async function getNewPosts(req: Request, res: Response) {
-  const posts = await postService.getNewPosts();
+  const userId = res.locals.user.id;
+  const posts = await postService.getNewPosts(userId);
   res.json(posts);
 }
