+++ back-end/src/services/postService.ts
@@ -13,23 +13,23 @@ async function publishPost(postId: number, userId: number) {
   return await postRepository.publish(postId);
 }
 
-async function getNewPosts() {
+async function getNewPosts(userId: number) {
   const posts = await postRepository.findAll();
   return await Promise.all(
     posts.map(async (post) => {
-      const { id: postId, userId } = post;
+      const { id: postId } = post;
       const likes = await likeService.getLikes({ postId, userId });
       return { ...post, likes };
     })
   );
 }
 
-async function getBestPosts() {
+async function getBestPosts(userId: number) {
   const posts = await postRepository.findLastWeek();
   posts.sort((a, b) => b._count.Like - a._count.Like);
   return await Promise.all(
     posts.slice(0, 10).map(async (post) => {
-      const { id: postId, userId } = post;
+      const { id: postId } = post;
       const likes = await likeService.getLikes({ postId, userId });
       delete post._count;
       return { ...post, likes };
@@ -41,7 +41,7 @@ async function getUserPosts(userId: number) {
   const posts = await postRepository.findUser(userId);
   return await Promise.all(
     posts.map(async (post) => {
-      const { id: postId, userId } = post;
+      const { id: postId } = post;
       const likes = await likeService.getLikes({ postId, userId });
       return { ...post, likes };
     })
@@ -52,7 +52,7 @@ async function getLikedPosts(userId: number) {
   const posts = await postRepository.findLiked(userId);
   return await Promise.all(
     posts.map(async (post) => {
-      const { id: postId, userId } = post.post;
+      const { id: postId } = post.post;
       const likes = await likeService.getLikes({ postId, userId });
       return { ...post.post, likes };
     })
