+++ front-end/src/components/Tree.js
@@ -8,18 +8,22 @@ export default class Tree extends Component {
   dir = true;
   width = this.props.width || window.screen.availWidth;
   height = this.props.height || window.screen.availHeight - 150;
+  mouseX=0;
 
   setup = (p5, parent) => {
     p5.createCanvas(this.width, this.height).parent(parent);
   };
   draw = (p5) => {
+    if(p5.mouseX < this.width && p5.mouseX > 0 && p5.mouseY < this.height && p5.mouseY > 0) {
+      this.mouseX = p5.mouseX;
+    }
     p5.background(0);
     p5.frameRate(20);
 
-    p5.stroke(p5.color(this.props.color || "#FFFFFF"));
+    p5.stroke(p5.color(this.props.color));
     p5.strokeWeight(this.width / 70);
     // Let's pick an angle 0 to 90 degrees based on the mouse position
-    let a = (p5.mouseX / this.width) * 90;
+    let a = (this.mouseX / this.width) * 90;
     // Convert it to radians
     this.theta = p5.radians(a);
     // Start the tree from the bottom of the screen
@@ -30,7 +34,7 @@ export default class Tree extends Component {
     p5.translate(0, -this.width / 3);
     // Start the recursive branching!
     this.branch(p5, this.width / 4);
-    p5.fill(p5.color(this.props.color || "#FFFFFF"));
+    p5.fill(p5.color(this.props.color));
     p5.ellipse(this.x, this.y, 20, 20);
     if (this.dir) {
       this.x += 10;
