+++ front-end/src/pages/Discover.js
@@ -11,9 +11,14 @@ function Discover() {
   const [loading, setLoading] = useState(true);
   const URL = "https://rys-gaiasenses.herokuapp.com/post/new";
   //const URL = "http://localhost:5000/post/new";
+  
 
   const [posts, setPosts] = useState([]);
   const { userData } = useContext(UserContext);
+  const [isShowingAlert, setShowingAlert] = useState(true);
+  setTimeout(() => {
+    setShowingAlert(false);
+  }, 2000);
 
   useEffect(() => {
     axios
@@ -33,6 +38,9 @@ function Discover() {
 
   return (
     <Main>
+      <Popup display={isShowingAlert}>
+          <p>Discover! Find out our newest posts!</p>
+      </Popup>
       <ConfigBar />
       {loading ? (
         <div className="noPost"><Circles color="#00bcd4" /></div>
@@ -59,5 +67,26 @@ const Main = styled.main`
     display: flex;
     justify-content: center;
     align-items: center;
+  }
+`;
+
+const Popup = styled.div`
+  position: fixed;
+  top: 100px;
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  opacity: ${(props) => (props.display ? 1 : 0)};
+  transition: all 250ms linear 2s;
+  z-index: 20;
+
+  p {
+    background-color: rgba(255, 255, 255, 0.5);
+    font-size: calc(1vw + 10px);
+    text-align: center;
+    max-width: 50%;
+    padding: 1vw;
+    border-radius: 1vw;
   }
 `;
